Add files via upload
[1]https://www.systemicpeace.org/inscrdata.html
The Integrated Network for Societal Conflict Research (INSCR) was established to coordinate and integrate information resources produced and used by the Center for Systemic Peace. The following data resources were prepared by researchers associated with the Center for Systemic Peace and are generated and/or compiled using open source information. These resources are made available as a service to the research community. All CSP/INSCR data resources have been cross-checked with other data resources to ensure, as far as possible, that the information recorded is accurate, reliable, and comprehensive.